+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- add support for 128 and 256 bit integers https://github.com/plausible/ecto_ch/pull/181
+
 ## 0.3.8 (2024-07-01)
 
 - add DateTime64 precision https://github.com/plausible/ecto_ch/pull/179