+++ CHANGELOG.md
@@ -40,7 +40,7 @@
 
 ## 0.3.0 (2023-09-13)
 
-- ensure all colums are for the same table in `ecto.ch.schema` mix task https://github.com/plausible/ecto_ch/pull/118
+- ensure all columns are for the same table in `ecto.ch.schema` mix task https://github.com/plausible/ecto_ch/pull/118
 - switch from `CREATE INDEX` to `ALTER TABLE ... ADD INDEX` syntax for indexes https://github.com/plausible/ecto_ch/pull/120
 - raise on `CREATE INDEX CONCURRENTLY` https://github.com/plausible/ecto_ch/pull/121
 - switch from `DROP INDEX` to `ALTER TABLE ... DROP INDEX` syntax for indexes https://github.com/plausible/ecto_ch/pull/122